+++ CHANGELOG.md
@@ -14,6 +14,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 
 - updated dependencies and template
+- corrected and clarified the documentation and parameter descriptions of the
+  `Validate Entities` and `Validate Knowledge Graph` plugins
+
+### Fixed
+
+- Validate Entities task no longer converts non-ASCII characters to unicode escape sequences
+  when writing valid JSON objects to the target dataset
 
 ## [1.3.0] 2026-08-19